Modification- dev-amitha - remove all the unwanted things and fix
